test: split providers-github monolith into per-surface facade tests (refs #4084)
Migrate the 54 contract-shaped GitHubProvider facade cases out of the
1583-line root monolith tests/providers-github.test.js into the existing
tests/providers/github/ tier, grouped by API surface:

- provider-issues.test.js (listIssues/getEpics, getEpic, getTicket,
  getTicketDependencies, createTicket, updateTicket)
- provider-comments.test.js (postComment)
- provider-prs.test.js (createPullRequest)
- provider-branch-protection.test.js (get/setBranchProtection, branchExists)
- provider-merge-methods.test.js (get/setMergeMethods)
- provider-labels.test.js (ensureLabels + Story #2018 post-loop verification)
- provider-projects.test.js (ensureProjectFields, isInsufficientScopes,
  ensureStatusField, ensureProjectViews)
- provider-core.test.js (construction + cross-cutting error handling)

The shared mocked-fetch / gh-exec setup helpers (makeGh, createGhExec,
createRouteMock, createTestProvider) move into a single _helpers.js so the
slices stay DRY. Every assertion is preserved verbatim; the GitHub contract
slice now runs and times independently, and the 1583-line review surface is
retired. Parity confirmed: 54 tests / 21 suites, all green, no live network.
